[v0.2.8-pre.1] Package renamed
