+++ Editor/Files/EditorDatabaseAsset.cs
@@ -6,17 +6,17 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Depra.Assets.Delegates;
-using Depra.Assets.Exceptions;
-using Depra.Assets.Extensions;
-using Depra.Assets.Files;
-using Depra.Assets.Files.Database;
-using Depra.Assets.ValueObjects;
+using Depra.Asset.Delegates;
+using Depra.Asset.Exceptions;
+using Depra.Asset.Extensions;
+using Depra.Asset.Files;
+using Depra.Asset.Files.Database;
+using Depra.Asset.ValueObjects;
 using UnityEditor;
 using UnityEngine;
 using Object = UnityEngine.Object;
 
-namespace Depra.Assets.Editor.Files
+namespace Depra.Asset.Editor.Files
 {
 	public sealed class EditorDatabaseAsset<TAsset> : IAssetFile<TAsset>, IDisposable where TAsset : ScriptableObject
 	{
