+++ Plugins/Depra.Asset.dll.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 12144cd499077c94f82093417464f995
+guid: 94dc0d35dd433de4cbd96cd7e3c3f342
 PluginImporter:
   externalObjects: {}
   serializedVersion: 2
@@ -7,7 +7,7 @@ PluginImporter:
   executionOrder: {}
   defineConstraints: []
   isPreloaded: 0
-  isOverridable: 1
+  isOverridable: 0
   isExplicitlyReferenced: 0
   validateReferences: 1
   platformData:
