+++ Runtime/Depra.Asset.Unity.asmdef
@@ -1,12 +1,8 @@
 {
-    "name": "Depra.Asset.Editor",
-    "rootNamespace": "Depra.Assets",
-    "references": [
-        "GUID:ce2ab5ef472b35f498b9d47a89be9396"
-    ],
-    "includePlatforms": [
-        "Editor"
-    ],
+    "name": "Depra.Asset.Unity",
+    "rootNamespace": "Depra.Asset",
+    "references": [],
+    "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
     "overrideReferences": true,
