+++ Runtime/Extensions/DirectoryInfoExtensions.cs
@@ -4,7 +4,7 @@
 using System.IO;
 using System.Runtime.CompilerServices;
 
-namespace Depra.Assets.Extensions
+namespace Depra.Asset.Extensions
 {
 	public static class DirectoryInfoExtensions
 	{
