+++ Runtime/Files.Bundles/Depra.Assets.Bundles.asmdef.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 007f974573e04f5cbfab7ed51f223fdc
+timeCreated: 1719335572
