+++ Runtime/Files.Bundles/Extensions/AssetBundleCreateRequestExtensions.cs
@@ -6,7 +6,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using UnityEngine;
 
-namespace Depra.Assets.Files.Bundles.Extensions
+namespace Depra.Asset.Files.Bundles.Extensions
 {
 	internal static class AssetBundleCreateRequestExtensions
 	{
