+++ Runtime/Files.Bundles/Sources/AssetBundleFromFile.cs
@@ -4,12 +4,12 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Depra.Assets.Exceptions;
-using Depra.Assets.Files.Bundles.Extensions;
-using Depra.Assets.ValueObjects;
+using Depra.Asset.Exceptions;
+using Depra.Asset.Files.Bundles.Extensions;
+using Depra.Asset.ValueObjects;
 using UnityEngine;
 
-namespace Depra.Assets.Files.Bundles.Sources
+namespace Depra.Asset.Files.Bundles.Sources
 {
 	public readonly struct AssetBundleFromFile : IAssetBundleSource
 	{
