+++ Runtime/Files.Bundles/Sources/AssetBundleSize.cs
@@ -5,11 +5,11 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using Depra.Assets.ValueObjects;
+using Depra.Asset.ValueObjects;
 using UnityEngine;
 using Object = UnityEngine.Object;
 
-namespace Depra.Assets.Files.Bundles.Sources
+namespace Depra.Asset.Files.Bundles.Sources
 {
 	public static class AssetBundleSize
 	{
