+++ Runtime/Files.Bundles/Sources/IAssetBundleSource.cs
@@ -4,10 +4,10 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Depra.Assets.ValueObjects;
+using Depra.Asset.ValueObjects;
 using UnityEngine;
 
-namespace Depra.Assets.Files.Bundles.Sources
+namespace Depra.Asset.Files.Bundles.Sources
 {
 	public interface IAssetBundleSource
 	{
