+++ Runtime/Files.Resource/ResourcesReference.cs
@@ -5,7 +5,7 @@ using System;
 using UnityEngine;
 using Object = UnityEngine.Object;
 
-namespace Depra.Assets.Files.Resource
+namespace Depra.Asset.Files.Resource
 {
 	[Serializable]
 	public sealed class ResourcesReference<TAsset> : ResourcesReference where TAsset : Object { }
