+++ Runtime/Internal/AssetTypes.cs
@@ -1,7 +1,7 @@
 ﻿// SPDX-License-Identifier: Apache-2.0
 // © 2023-2024 Nikolay Melnikov <n.melnikov@depra.org>
 
-namespace Depra.Assets.Common
+namespace Depra.Asset.Internal
 {
 	public static class AssetTypes
 	{
