+++ Runtime/Internal/AssetTypes.cs.meta
