+++ Runtime/Internal/Module.cs
@@ -1,11 +1,11 @@
 ﻿// SPDX-License-Identifier: Apache-2.0
 // © 2023-2024 Nikolay Melnikov <n.melnikov@depra.org>
 
-namespace Depra.Assets.Common
+namespace Depra.Asset.Internal
 {
 	internal static class Module
 	{
-		public const string MODULE_NAME = nameof(Assets);
+		public const string MODULE_NAME = nameof(Asset);
 		public const string FRAMEWORK_NAME = nameof(Depra);
 
 		internal const int DEFAULT_ORDER = 52;
