+++ Runtime/Properties/AssemblyInfo.cs
@@ -1,9 +1,10 @@
 ﻿using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("Depra.Assets.Editor")]
+[assembly: InternalsVisibleTo("Depra.Asset.Editor")]
+[assembly: InternalsVisibleTo("Depra.Asset.Bundles")]
 
-[assembly: InternalsVisibleTo("Depra.Assets.Tests.EditMode")]
-[assembly: InternalsVisibleTo("Depra.Assets.Tests.PlayMode")]
+[assembly: InternalsVisibleTo("Depra.Asset.Tests.EditMode")]
+[assembly: InternalsVisibleTo("Depra.Asset.Tests.PlayMode")]
 
-[assembly: InternalsVisibleTo("Depra.Assets.Addressable.Runtime")]
-[assembly: InternalsVisibleTo("Depra.Assets.Addressable.Tests")]
+[assembly: InternalsVisibleTo("Depra.Asset.Addressable.Runtime")]
+[assembly: InternalsVisibleTo("Depra.Asset.Addressable.Tests")]
